Enable using Unix sockets to communicate with containers (#48)
- Create Unix socket server that runs inside containers.
- Abstract container communication from container creation.
- Each container now has a special host folder for its "stuff"
- Using socket has ~6x reduction (~2-3 ms => 0.5 ms) in E2E platform
overhead

How it works:
- Worker passes a environment var to the container indicating the socket
path it must open.
- That folder is mounted from host container folder to that path
- Worker listens on that to be created.
- Uses a 2-message system to invoke.
- First is a `SendMessage` with the first a u64 `Command::Invoke` and
the len in bytes of the str arguments
  - Then the arguments are sent and deserialized by the worker
- Same system of sending message metadata first used to return results 

TODOs:
- Add details to benchmark CSVs
- Push GPU containers that use Unix
- p99 results of some sort

This change is pretty much done, but I found a related problem. The
worker runs into a scaling issue (or something) at high levels of client
calls. Such that when running ~40 clients with high frequency
invocations, using the `scaling` part of load gen and the simple `hello`
image, the worker seems to stop responding. Not related to this issue I
think, as it happens to the old HTTP client too