Re-write characteristics map (#53)
A greatly simplified but easier to use and extend characteristics map.
Each func has an array of values in a dashmap, so we can access any
value with only one RW lock access.
This also lets us read or write to several values at once with one lock.

Extensible, so a custom `enum` can be passed. Each entry in the enum
(see `Chars`) is given four entries in the, one for each `Value`.
Cmap tracks Min, Max, (running) Avg, and Latest value.
These are all updated when the entry is updated, and any can be
requested.

Did not implement for this phase, but it will be easy to export all the
data on a per-func basis thanks to the new flat layout.